+++ VTC2/character_manager.cpp
@@ -37,7 +37,15 @@ void CharacterManager::Update(double timeTotal, double timeDelta)
 void CharacterManager::Render(Graphics* graphics)
 {
 	for (Character* character : allCharacters) {
-		graphics->DrawCircle(character->position, 4, D2D1::ColorF(character->affiliationId, 0, 1-character->affiliationId, character->alive?1:0.2));
-		graphics->DrawTextVTC(std::to_string(character->partyId), character->position, D2D1::ColorF(1, 1, 1, character->alive ? 1 : 0.2));
+		graphics->DrawCircle(
+			character->position
+			, 4
+			, D2D1::ColorF((float)character->affiliationId, 0, (float)1-character->affiliationId, character->alive ? 1 : 0.2f)
+		);
+		graphics->DrawTextVTC(
+			std::to_string(character->partyId)
+			, character->position
+			, D2D1::ColorF(1, 1, 1, character->alive ? 1 : 0.2f)
+		);
 	}
 }
